test: 100% coverage for rate-limit.ts and fix security hotspots
- Reached 100% statement coverage for `app-next-directory/src/utils/rate-limit.ts`.
- Consolidated IP extraction and validation logic into a new shared utility `src/utils/ip-utils.ts` using `validator.isIP`.
- Refactored `rate-limit.ts`, `auth.ts`, `logger.ts`, and `src/lib/rate-limit.ts` to use the shared utility, resolving SonarCloud duplication and security hotspots.
- Pinned `google-github-actions/run-gemini-cli` to `@v0` in workflow files to fix CI resolution errors.
- Added comprehensive unit tests in `src/utils/__tests__/rate-limit.test.ts`.
- Fixed Redis lazy initialization and 'open handle' leaks in tests.
